+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 
 ## [0.2.1]
 - Add documentation
+- Fix typing
+- Remove useless GameList serialisation with `pickle`
 
 ## [0.2.0]
 - Add icon